Added docstrings to properties. Began adding energy method. Some docstrings run over their lines, need to fix

+++ source/stability.cpp
@@ -8,7 +8,7 @@
 #include "armadillo"
 #define ARMA_NO_DEBUG
 
-double HFStability::HEG::energy(unsigned int state)
+double HFStability::HEG::energy(long long unsigned int state)
 {
 	double energy;
 	double kin = 0; //kinetic energy
